requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
python-dotenv
celery[redis]